requirements.txt
python-telegram-bot==21.11
beautifulsoup4==4.12.3
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.9
PyJWT==2.10.1
httpx==0.27.2
